refactor: refactor backend types and improve documentation
