+++ frontend/.gitignore
@@ -13,6 +13,9 @@ dev-dist
 dist-ssr
 *.local
 
+# Auto generated
+src/bindings/*
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
